LPS-146764 Build service

+++ modules/apps/depot/depot-service/src/main/java/com/liferay/depot/service/persistence/impl/DepotAppCustomizationPersistenceImpl.java
@@ -23,6 +23,7 @@ import com.liferay.depot.service.persistence.DepotAppCustomizationPersistence;
 import com.liferay.depot.service.persistence.DepotAppCustomizationUtil;
 import com.liferay.depot.service.persistence.impl.constants.DepotPersistenceConstants;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.kernel.change.tracking.CTColumnResolutionType;
 import com.liferay.portal.kernel.configuration.Configuration;
 import com.liferay.portal.kernel.dao.orm.EntityCache;
 import com.liferay.portal.kernel.dao.orm.FinderCache;
@@ -36,6 +37,7 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.service.persistence.BasePersistence;
+import com.liferay.portal.kernel.service.persistence.change.tracking.helper.CTPersistenceHelper;
 import com.liferay.portal.kernel.service.persistence.impl.BasePersistenceImpl;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
@@ -49,7 +51,12 @@ import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationHandler;
 
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.EnumMap;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -173,18 +180,21 @@ public class DepotAppCustomizationPersistenceImpl
 		OrderByComparator<DepotAppCustomization> orderByComparator,
 		boolean useFinderCache) {
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
+
 		FinderPath finderPath = null;
 		Object[] finderArgs = null;
 
 		if ((start == QueryUtil.ALL_POS) && (end == QueryUtil.ALL_POS) &&
 			(orderByComparator == null)) {
 
-			if (useFinderCache) {
+			if (useFinderCache && productionMode) {
 				finderPath = _finderPathWithoutPaginationFindByDepotEntryId;
 				finderArgs = new Object[] {depotEntryId};
 			}
 		}
-		else if (useFinderCache) {
+		else if (useFinderCache && productionMode) {
 			finderPath = _finderPathWithPaginationFindByDepotEntryId;
 			finderArgs = new Object[] {
 				depotEntryId, start, end, orderByComparator
@@ -193,7 +203,7 @@ public class DepotAppCustomizationPersistenceImpl
 
 		List<DepotAppCustomization> list = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			list = (List<DepotAppCustomization>)finderCache.getResult(
 				finderPath, finderArgs);
 
@@ -251,7 +261,7 @@ public class DepotAppCustomizationPersistenceImpl
 
 				cacheResult(list);
 
-				if (useFinderCache) {
+				if (useFinderCache && productionMode) {
 					finderCache.putResult(finderPath, finderArgs, list);
 				}
 			}
@@ -561,11 +571,21 @@ public class DepotAppCustomizationPersistenceImpl
 	 */
 	@Override
 	public int countByDepotEntryId(long depotEntryId) {
-		FinderPath finderPath = _finderPathCountByDepotEntryId;
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
 
-		Object[] finderArgs = new Object[] {depotEntryId};
+		FinderPath finderPath = null;
+		Object[] finderArgs = null;
 
-		Long count = (Long)finderCache.getResult(finderPath, finderArgs);
+		Long count = null;
+
+		if (productionMode) {
+			finderPath = _finderPathCountByDepotEntryId;
+
+			finderArgs = new Object[] {depotEntryId};
+
+			count = (Long)finderCache.getResult(finderPath, finderArgs);
+		}
 
 		if (count == null) {
 			StringBundler sb = new StringBundler(2);
@@ -589,7 +609,9 @@ public class DepotAppCustomizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(finderPath, finderArgs, count);
+				if (productionMode) {
+					finderCache.putResult(finderPath, finderArgs, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -672,15 +694,18 @@ public class DepotAppCustomizationPersistenceImpl
 	public DepotAppCustomization fetchByD_E(
 		long depotEntryId, boolean enabled, boolean useFinderCache) {
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
+
 		Object[] finderArgs = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			finderArgs = new Object[] {depotEntryId, enabled};
 		}
 
 		Object result = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			result = finderCache.getResult(_finderPathFetchByD_E, finderArgs);
 		}
 
@@ -722,7 +747,7 @@ public class DepotAppCustomizationPersistenceImpl
 				List<DepotAppCustomization> list = query.list();
 
 				if (list.isEmpty()) {
-					if (useFinderCache) {
+					if (useFinderCache && productionMode) {
 						finderCache.putResult(
 							_finderPathFetchByD_E, finderArgs, list);
 					}
@@ -732,7 +757,7 @@ public class DepotAppCustomizationPersistenceImpl
 						Collections.sort(list, Collections.reverseOrder());
 
 						if (_log.isWarnEnabled()) {
-							if (!useFinderCache) {
+							if (!productionMode || !useFinderCache) {
 								finderArgs = new Object[] {
 									depotEntryId, enabled
 								};
@@ -794,11 +819,21 @@ public class DepotAppCustomizationPersistenceImpl
 	 */
 	@Override
 	public int countByD_E(long depotEntryId, boolean enabled) {
-		FinderPath finderPath = _finderPathCountByD_E;
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
 
-		Object[] finderArgs = new Object[] {depotEntryId, enabled};
+		FinderPath finderPath = null;
+		Object[] finderArgs = null;
 
-		Long count = (Long)finderCache.getResult(finderPath, finderArgs);
+		Long count = null;
+
+		if (productionMode) {
+			finderPath = _finderPathCountByD_E;
+
+			finderArgs = new Object[] {depotEntryId, enabled};
+
+			count = (Long)finderCache.getResult(finderPath, finderArgs);
+		}
 
 		if (count == null) {
 			StringBundler sb = new StringBundler(3);
@@ -826,7 +861,9 @@ public class DepotAppCustomizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(finderPath, finderArgs, count);
+				if (productionMode) {
+					finderCache.putResult(finderPath, finderArgs, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -914,15 +951,18 @@ public class DepotAppCustomizationPersistenceImpl
 
 		portletId = Objects.toString(portletId, "");
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
+
 		Object[] finderArgs = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			finderArgs = new Object[] {depotEntryId, portletId};
 		}
 
 		Object result = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			result = finderCache.getResult(_finderPathFetchByD_P, finderArgs);
 		}
 
@@ -976,7 +1016,7 @@ public class DepotAppCustomizationPersistenceImpl
 				List<DepotAppCustomization> list = query.list();
 
 				if (list.isEmpty()) {
-					if (useFinderCache) {
+					if (useFinderCache && productionMode) {
 						finderCache.putResult(
 							_finderPathFetchByD_P, finderArgs, list);
 					}
@@ -1034,11 +1074,21 @@ public class DepotAppCustomizationPersistenceImpl
 	public int countByD_P(long depotEntryId, String portletId) {
 		portletId = Objects.toString(portletId, "");
 
-		FinderPath finderPath = _finderPathCountByD_P;
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
 
-		Object[] finderArgs = new Object[] {depotEntryId, portletId};
+		FinderPath finderPath = null;
+		Object[] finderArgs = null;
 
-		Long count = (Long)finderCache.getResult(finderPath, finderArgs);
+		Long count = null;
+
+		if (productionMode) {
+			finderPath = _finderPathCountByD_P;
+
+			finderArgs = new Object[] {depotEntryId, portletId};
+
+			count = (Long)finderCache.getResult(finderPath, finderArgs);
+		}
 
 		if (count == null) {
 			StringBundler sb = new StringBundler(3);
@@ -1077,7 +1127,9 @@ public class DepotAppCustomizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(finderPath, finderArgs, count);
+				if (productionMode) {
+					finderCache.putResult(finderPath, finderArgs, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -1115,6 +1167,10 @@ public class DepotAppCustomizationPersistenceImpl
 	 */
 	@Override
 	public void cacheResult(DepotAppCustomization depotAppCustomization) {
+		if (depotAppCustomization.getCtCollectionId() != 0) {
+			return;
+		}
+
 		entityCache.putResult(
 			DepotAppCustomizationImpl.class,
 			depotAppCustomization.getPrimaryKey(), depotAppCustomization);
@@ -1157,6 +1213,10 @@ public class DepotAppCustomizationPersistenceImpl
 
 		for (DepotAppCustomization depotAppCustomization :
 				depotAppCustomizations) {
+
+			if (depotAppCustomization.getCtCollectionId() != 0) {
+				continue;
+			}
 
 			if (entityCache.getResult(
 					DepotAppCustomizationImpl.class,
@@ -1326,7 +1386,9 @@ public class DepotAppCustomizationPersistenceImpl
 					depotAppCustomization.getPrimaryKeyObj());
 			}
 
-			if (depotAppCustomization != null) {
+			if ((depotAppCustomization != null) &&
+				ctPersistenceHelper.isRemove(depotAppCustomization)) {
+
 				session.delete(depotAppCustomization);
 			}
 		}
@@ -1377,7 +1439,13 @@ public class DepotAppCustomizationPersistenceImpl
 		try {
 			session = openSession();
 
-			if (isNew) {
+			if (ctPersistenceHelper.isInsert(depotAppCustomization)) {
+				if (!isNew) {
+					session.evict(
+						DepotAppCustomizationImpl.class,
+						depotAppCustomization.getPrimaryKeyObj());
+				}
+
 				session.save(depotAppCustomization);
 			}
 			else {
@@ -1390,6 +1458,16 @@ public class DepotAppCustomizationPersistenceImpl
 		}
 		finally {
 			closeSession(session);
+		}
+
+		if (depotAppCustomization.getCtCollectionId() != 0) {
+			if (isNew) {
+				depotAppCustomization.setNew(false);
+			}
+
+			depotAppCustomization.resetOriginalValues();
+
+			return depotAppCustomization;
 		}
 
 		entityCache.putResult(
@@ -1450,6 +1528,42 @@ public class DepotAppCustomizationPersistenceImpl
 	/**
 	 * Returns the depot app customization with the primary key or returns <code>null</code> if it could not be found.
 	 *
+	 * @param primaryKey the primary key of the depot app customization
+	 * @return the depot app customization, or <code>null</code> if a depot app customization with the primary key could not be found
+	 */
+	@Override
+	public DepotAppCustomization fetchByPrimaryKey(Serializable primaryKey) {
+		if (ctPersistenceHelper.isProductionMode(DepotAppCustomization.class)) {
+			return super.fetchByPrimaryKey(primaryKey);
+		}
+
+		DepotAppCustomization depotAppCustomization = null;
+
+		Session session = null;
+
+		try {
+			session = openSession();
+
+			depotAppCustomization = (DepotAppCustomization)session.get(
+				DepotAppCustomizationImpl.class, primaryKey);
+
+			if (depotAppCustomization != null) {
+				cacheResult(depotAppCustomization);
+			}
+		}
+		catch (Exception exception) {
+			throw processException(exception);
+		}
+		finally {
+			closeSession(session);
+		}
+
+		return depotAppCustomization;
+	}
+
+	/**
+	 * Returns the depot app customization with the primary key or returns <code>null</code> if it could not be found.
+	 *
 	 * @param depotAppCustomizationId the primary key of the depot app customization
 	 * @return the depot app customization, or <code>null</code> if a depot app customization with the primary key could not be found
 	 */
@@ -1458,6 +1572,102 @@ public class DepotAppCustomizationPersistenceImpl
 		long depotAppCustomizationId) {
 
 		return fetchByPrimaryKey((Serializable)depotAppCustomizationId);
+	}
+
+	@Override
+	public Map<Serializable, DepotAppCustomization> fetchByPrimaryKeys(
+		Set<Serializable> primaryKeys) {
+
+		if (ctPersistenceHelper.isProductionMode(DepotAppCustomization.class)) {
+			return super.fetchByPrimaryKeys(primaryKeys);
+		}
+
+		if (primaryKeys.isEmpty()) {
+			return Collections.emptyMap();
+		}
+
+		Map<Serializable, DepotAppCustomization> map =
+			new HashMap<Serializable, DepotAppCustomization>();
+
+		if (primaryKeys.size() == 1) {
+			Iterator<Serializable> iterator = primaryKeys.iterator();
+
+			Serializable primaryKey = iterator.next();
+
+			DepotAppCustomization depotAppCustomization = fetchByPrimaryKey(
+				primaryKey);
+
+			if (depotAppCustomization != null) {
+				map.put(primaryKey, depotAppCustomization);
+			}
+
+			return map;
+		}
+
+		if ((databaseInMaxParameters > 0) &&
+			(primaryKeys.size() > databaseInMaxParameters)) {
+
+			Iterator<Serializable> iterator = primaryKeys.iterator();
+
+			while (iterator.hasNext()) {
+				Set<Serializable> page = new HashSet<>();
+
+				for (int i = 0;
+					 (i < databaseInMaxParameters) && iterator.hasNext(); i++) {
+
+					page.add(iterator.next());
+				}
+
+				map.putAll(fetchByPrimaryKeys(page));
+			}
+
+			return map;
+		}
+
+		StringBundler sb = new StringBundler((primaryKeys.size() * 2) + 1);
+
+		sb.append(getSelectSQL());
+		sb.append(" WHERE ");
+		sb.append(getPKDBName());
+		sb.append(" IN (");
+
+		for (Serializable primaryKey : primaryKeys) {
+			sb.append((long)primaryKey);
+
+			sb.append(",");
+		}
+
+		sb.setIndex(sb.index() - 1);
+
+		sb.append(")");
+
+		String sql = sb.toString();
+
+		Session session = null;
+
+		try {
+			session = openSession();
+
+			Query query = session.createQuery(sql);
+
+			for (DepotAppCustomization depotAppCustomization :
+					(List<DepotAppCustomization>)query.list()) {
+
+				map.put(
+					depotAppCustomization.getPrimaryKeyObj(),
+					depotAppCustomization);
+
+				cacheResult(depotAppCustomization);
+			}
+		}
+		catch (Exception exception) {
+			throw processException(exception);
+		}
+		finally {
+			closeSession(session);
+		}
+
+		return map;
 	}
 
 	/**
@@ -1525,25 +1735,28 @@ public class DepotAppCustomizationPersistenceImpl
 		OrderByComparator<DepotAppCustomization> orderByComparator,
 		boolean useFinderCache) {
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
+
 		FinderPath finderPath = null;
 		Object[] finderArgs = null;
 
 		if ((start == QueryUtil.ALL_POS) && (end == QueryUtil.ALL_POS) &&
 			(orderByComparator == null)) {
 
-			if (useFinderCache) {
+			if (useFinderCache && productionMode) {
 				finderPath = _finderPathWithoutPaginationFindAll;
 				finderArgs = FINDER_ARGS_EMPTY;
 			}
 		}
-		else if (useFinderCache) {
+		else if (useFinderCache && productionMode) {
 			finderPath = _finderPathWithPaginationFindAll;
 			finderArgs = new Object[] {start, end, orderByComparator};
 		}
 
 		List<DepotAppCustomization> list = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			list = (List<DepotAppCustomization>)finderCache.getResult(
 				finderPath, finderArgs);
 		}
@@ -1581,7 +1794,7 @@ public class DepotAppCustomizationPersistenceImpl
 
 				cacheResult(list);
 
-				if (useFinderCache) {
+				if (useFinderCache && productionMode) {
 					finderCache.putResult(finderPath, finderArgs, list);
 				}
 			}
@@ -1614,8 +1827,15 @@ public class DepotAppCustomizationPersistenceImpl
 	 */
 	@Override
 	public int countAll() {
-		Long count = (Long)finderCache.getResult(
-			_finderPathCountAll, FINDER_ARGS_EMPTY);
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			DepotAppCustomization.class);
+
+		Long count = null;
+
+		if (productionMode) {
+			count = (Long)finderCache.getResult(
+				_finderPathCountAll, FINDER_ARGS_EMPTY);
+		}
 
 		if (count == null) {
 			Session session = null;
@@ -1628,8 +1848,10 @@ public class DepotAppCustomizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(
-					_finderPathCountAll, FINDER_ARGS_EMPTY, count);
+				if (productionMode) {
+					finderCache.putResult(
+						_finderPathCountAll, FINDER_ARGS_EMPTY, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -1658,8 +1880,61 @@ public class DepotAppCustomizationPersistenceImpl
 	}
 
 	@Override
-	protected Map<String, Integer> getTableColumnsMap() {
+	public Set<String> getCTColumnNames(
+		CTColumnResolutionType ctColumnResolutionType) {
+
+		return _ctColumnNamesMap.getOrDefault(
+			ctColumnResolutionType, Collections.emptySet());
+	}
+
+	@Override
+	public List<String> getMappingTableNames() {
+		return _mappingTableNames;
+	}
+
+	@Override
+	public Map<String, Integer> getTableColumnsMap() {
 		return DepotAppCustomizationModelImpl.TABLE_COLUMNS_MAP;
+	}
+
+	@Override
+	public String getTableName() {
+		return "DepotAppCustomization";
+	}
+
+	@Override
+	public List<String[]> getUniqueIndexColumnNames() {
+		return _uniqueIndexColumnNames;
+	}
+
+	private static final Map<CTColumnResolutionType, Set<String>>
+		_ctColumnNamesMap = new EnumMap<CTColumnResolutionType, Set<String>>(
+			CTColumnResolutionType.class);
+	private static final List<String> _mappingTableNames =
+		new ArrayList<String>();
+	private static final List<String[]> _uniqueIndexColumnNames =
+		new ArrayList<String[]>();
+
+	static {
+		Set<String> ctControlColumnNames = new HashSet<String>();
+		Set<String> ctStrictColumnNames = new HashSet<String>();
+
+		ctControlColumnNames.add("mvccVersion");
+		ctControlColumnNames.add("ctCollectionId");
+		ctStrictColumnNames.add("companyId");
+		ctStrictColumnNames.add("depotEntryId");
+		ctStrictColumnNames.add("enabled");
+		ctStrictColumnNames.add("portletId");
+
+		_ctColumnNamesMap.put(
+			CTColumnResolutionType.CONTROL, ctControlColumnNames);
+		_ctColumnNamesMap.put(
+			CTColumnResolutionType.PK,
+			Collections.singleton("depotAppCustomizationId"));
+		_ctColumnNamesMap.put(
+			CTColumnResolutionType.STRICT, ctStrictColumnNames);
+
+		_uniqueIndexColumnNames.add(new String[] {"depotEntryId", "portletId"});
 	}
 
 	/**
@@ -1771,6 +2046,9 @@ public class DepotAppCustomizationPersistenceImpl
 	public void setSessionFactory(SessionFactory sessionFactory) {
 		super.setSessionFactory(sessionFactory);
 	}
+
+	@Reference
+	protected CTPersistenceHelper ctPersistenceHelper;
 
 	@Reference
 	protected EntityCache entityCache;
